+++ CHANGELOG.md
@@ -43,14 +43,17 @@ Thank you to all who have contributed!
 ### Deprecated
 
 ### Fixed
+- Fixes typing of scalar subqueries in the PartiQLSchemaInferencer. Note that usage of `SELECT *` in subqueries
+  is not fully supported. Please make sure to handle InferenceExceptions.
 - Fixes schema inferencer behavior for ORDER BY
+
 ### Removed
 
 ### Security
 
 ### Contributors
 Thank you to all who have contributed!
-- @<your-username>
+- @johnedquinn
 
 ## [0.13.1] - 2023-09-19